STOUT-FACTORY: Use random digit as a pass key to prevent false positive.
It promots a random digit (0-9) on external display only with main
display turned off. Operator must see the external display to know which
key to pass the task.

BUG=chrome-os-partner:13738
TEST=manual
After connecting the external display, the main display should be turned off
and a random number is prompted on theh external display for operator to
pass the task.

Change-Id: I7b389cea232d441632e0b84b25e6cba8542e739f
Original-Change-Id: I0209eeb3ea600924ed750e783104bb81658e4f36
Reviewed-on: https://gerrit.chromium.org/gerrit/33556
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>