STOUT-FACTORY: Unmatched components should fail to match bom
When we match bom using probed components, we should not only check the
extant components do match. There might be some components not listed
in component_db, and are not matched to what the bom specified.
This CL fix MatchComponentSpecWithData such that it will check all the
components of this bom can find a matched probed component.

BUG=chrome-os-partner:14176
TEST=plug a wrong wifi not listed in component_db
     check gooftool verify_hwid will fail.

Change-Id: I37e2171d47645a7a283076d2b747258deab3f9c8
Original-Change-Id: I3f7fd755870e478053f4a8936837588b87af3079
Reviewed-on: https://gerrit.chromium.org/gerrit/33542
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>