clear_gbb_flag: should exit nonzero when fails to clear gbb flags
When firmware write-protection is enabled (both sw and hw), and gbb
flags is not zero, clear_gbb_flag should exit with nonzero return code.
The root cause is because || operator will disable set -e in its operand.
Although "invoke" returns nonzero value when it fails to change gbb flags,
the shell continues to execute.
Also, at the end of the script, it should exit 1 since if gbb was already
cleared it should have returned earlier.

BUG=chrome-os-partner:15161
TEST=1.set gbb_flag to 0x39 and enable write-protect both sw and hw.
     2.sh -x clear_gbb_flags.sh;echo $?
     3.check the return value is 1

Change-Id: I69329e58761d54cf633c7e597ec83668c9d1248d
Reviewed-on: https://gerrit.chromium.org/gerrit/35384
Reviewed-by: Jon Salz <jsalz@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>