Add a test to update firmware.
It is used to handle the case:
When a machine is in a middle of factory flow and it needs an emergency
firmware update, we cannot update /usr/sbin/chromeos-firmwareupdate as
it is in the read-only partition. We utilize the fact that factory code is
in stateful partition and use the updated factory (w/ new
chromeos-firmwareupdate in factory payload) to update firmware.

BUG=chrome-os-partner:12863
TEST=manual

Change-Id: Ieda0b8a4a506975efa58572b403ffcf41e5ef665
Reviewed-on: https://gerrit.chromium.org/gerrit/35575
Commit-Ready: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Shuo-Peng Liao <deanliao@chromium.org>
(cherry picked from commit c7fa6f05c7b87f407d4ba316ecd7476b10b52db7)
Reviewed-on: https://gerrit.chromium.org/gerrit/35790
Reviewed-by: Jon Salz <jsalz@chromium.org>
Tested-by: Jon Salz <jsalz@chromium.org>