finalize: add gooftool debug logs to factory.log
Currently the DEBUG output of gooftool in finalize does not go to
factory.log. The cause is that in finalize we use -l console.log to
execute gooftool and directly output the DEBUG message to console.log.
We would like to see DEBUG logs in both console.log as it
executes, and see all the log in factory.log.
To fix this problem, we can remove the -l argumemt of gooftool in
finalize, so the stderr(already redirected to stdout) which contains the
DEBUG message will go to console.log immediately through tee in
gooftool wrapper. And the stderr and stdout log will go to factory.log.

BUG=none
TEST=run finalize test, can see gooftool DEBUG message in console as it
executes. Also check that DEBUG message goes to factory.log

Change-Id: If87db89941e15e1444ab3b9ed8bf61d8e7768e18
Reviewed-on: https://gerrit.chromium.org/gerrit/37521
Reviewed-by: Jon Salz <jsalz@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>