Add retries argument to FactoryTest
Give FactoryTest certain retries to run a test until it pass.
retries is the maximum number of retries allowed to pass the test.
If it's 0, then no retries are allowed (the usual case). If, for example,
iterations=60 and retries=2, then the test would be run up to 62 times
and could fail up to twice.

BUG=chrome-os-partner:18550
TEST=use iterations=3 and retries=2 to run LED test
   See it gets two retries to pass LED test for 3 times.

Change-Id: I6552e6ba14a4f9d1b87e2a3f1e2a4a477983db2c
Reviewed-on: https://gerrit.chromium.org/gerrit/47252
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>