Add support for taking note on a device
This adds an item 'take note' in the CrOS menu. The notes will show up
on the UI with a click on the yellow exclamation mark. If the last note
is declared 'critical', no test is allowed to run anymore and the notes
are display on the UI directly.

BUG=chrome-os-partner:18665
TEST=Manual

Original-Change-Id: I6fea4f8d788ae995f55f2c5ae6e86b2be531c3bf
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47719
(cherry picked from commit aabf9fd865b3c601c20d8b77f764c3371bf8be30)

Change-Id: I59dfa58d13820acaf0f7c0d2648c0764a0f6f8f1
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47963
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>