Give the .device_id file the precedence for obtaining device id.
Currently the device id is bundled with the mac address of the wifi
module. However, it is possible that the module returns a different
mac address during some specific tests. This CL fixed the problem
by always first find the device id from the .device_id file, which
is a cache writtened when it is obtained very first time. This
makes the device id stable unless it is re-imaged.

BUG=chrome-os-partner:18500
TEST=unit tests

Change-Id: Iea6ad5bc642a158eae81965b157e052f1143c8f6
Reviewed-on: https://gerrit.chromium.org/gerrit/48352
Reviewed-by: Chun-Ta Lin <itspeter@google.com>
Commit-Queue: Andy Cheng <andycheng@chromium.org>
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Tested-by: Andy Cheng <andycheng@chromium.org>