CHERRY-PICK: Minijack: Add event_seq and drop dargs in the Test table
The event_seq's have the correct order of Test happening. The start_time
and end_time are not correct when time shifted.

Also drop the dargs as it is not useful in the Table table.

BUG=chrome-os-partner:19242
TEST=Manual
Run Minijack on the build log and check the new tags are added.

Change-Id: If87b3502b2a0ce902d45154e478f6907f609476f
Original-Change-Id: I1bdbe163fee0368a8995fcb0c1764eaf4fa3bdd4
Reviewed-on: https://gerrit.chromium.org/gerrit/55754
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>