CHERRY-PICK: Stop processing remaining when abort in EventLogWatcher
When calling StopWatchThread() to abort the EventLogWatcher, so far it needs to
wait all the event log processed. It takes very long time in the Minijack case,
as there may be a lot of unprocessed event log. As a result, stopping shopfloord
(which launches/stops Minijack) also takes a lot of time.

This change makes it skip the remaining in every num_log_per_callback. So Minijack
can be terminated quickly.

The behavior of Goofy keeps the same as its num_log_per_callback is infinity.

BUG=chrome-os-partner:18799
TEST=Manaul
Run Minijack to process a hugh number of log, press Ctrl-C to abort it, check it
terminated in a short time.

Change-Id: I17604f3bc530e14c397a00b844c2bc1c82f7c454
Original-Change-Id: I90f8377d25cecf9368e2a598bf05ccdcd59bdb27
Reviewed-on: https://gerrit.chromium.org/gerrit/55707
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>