CHERRY-PICK: Enable clean wakealarm extension for suspend resume test
As the original method of pushing out the wakealarm required clearing it
there was a race condition with the actual suspend that could lead to
a DUT being stuck in suspend. This leverages a new method of pushing out
the wakealarm that keeps it active to avoid this race.

This is dependent on the kernel CL at
https://gerrit.chromium.org/gerrit/#/c/50527/

BUG=chromium:239217
TEST=Manual, run test on DUTs.

Change-Id: Idf81ac22e7de3b56afe20dd9d66107a61add6734
Original-Change-Id: I2059c2bfc09a7b131451e0b725d19d4ca7ce4567
Reviewed-on: https://gerrit.chromium.org/gerrit/56794
Commit-Queue: Jon Salz <jsalz@chromium.org>
Tested-by: Jon Salz <jsalz@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>