Prompt for Ethernet if running netboot firmware
If we are running netboot firmware, show an error message on goofy
startup. This prompts the user to plug in Ethernet dongle and reboot to
re-image. This way if the Ethernet is not connected, netboot firmware
boots internal disk and the user sees the prompt.

BUG=chrome-os-partner:19891
TEST=On Spring, do flash_netboot and see the prompt after rebooting.

Original-Change-Id: I5c45b689849c2c1e6d43650350efaa13385777a1
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57435
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
(cherry picked from commit 9bd4f779eb692370b522a6dc8ab1b2a74a8fe81b)

Change-Id: I7d74c1b210fbdf0d964b656f3a8f0beb03cc3de2
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57600