CHERRY-PICK: RMA: Unpack the HWID bundle when needed
The HWID decoding feature of the simple RMA shopfloor server requires
access to the HWID HWDB, however this HWDB may not be unpacked in
the bundle, so we need to unpack it when we need it.

BUG=chrome-os-partner:21210
TEST=Manual, run simple RMA shopfloor, watch it unpack a HWDB from
a HWID bundle.

Original-Change-Id: Iec0ef3bc335801384fe85a2723f2b03d263039c4
(cherry picked from commit 459c789acd5bb0bb1570b85d30c74b9ea53751e9)

Change-Id: I1f605eb39f4330d98524a317a8bb104444164baf
Reviewed-on: https://gerrit.chromium.org/gerrit/62888
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>