serial_utils: open a port only when it is not open yet
When using serial.Serial(), the port is immediately open when the
object is created. On the other hand, when the port is None at
object creation, a successive open() is needed to open the port.

In this patch, the open() is invoked only when the port is not
open yet at object creation time.

BUG=chrome-os-partner:23254
TEST=Manual test with an arduino board.

Change-Id: I626e8a3afb2a93b4967e7c57bc523074c9a74c36
Previous-Reviewed-on: https://chromium-review.googlesource.com/173075
(cherry picked from commit 897b7fc947e52d59ebe6e4d2c529631f3205d3a2)
Reviewed-on: https://chromium-review.googlesource.com/176105
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>