CHERRY-PICK: Add touchscreen uniformity test
This test checks that the raw reference and delta information from
a touchscreen controller falls in a specified range. It is intended
to be run during run-in without a fixture or operator.

Note that the max and min values may need adjusting based on the
model being tested.

BUG=chrome-os-partner:23316
TEST=Manual. Run uniformity test on a device with an Atmel 1664s
touch controller using the example test_list configuration in the
docstring.

Change-Id: I3e6b4ddd5c33e00b4bdab27ec1f079aaa9917360
Original-Change-Id: Ie5273cd9869d6fd7ec02c3e87824a8a9fce405f1
Reviewed-on: https://chromium-review.googlesource.com/174640
Reviewed-by: Dave Parker <dparker@chromium.org>
Commit-Queue: Dave Parker <dparker@chromium.org>
Tested-by: Dave Parker <dparker@chromium.org>