CHERRY-PICK: Make the test check dongle after the operator has confirmed to start the test, not before.
In the audio loop test, if auto start is false, the test requires the
operator to press the "s" key to start the test. By now, the test checks
dongle status before that, so it will fail immediately if the dongle
status is wrong. This is not the intended result. We should give it a
chance to inform the operator to plug in/out the dongle.

Original-Change-Id: I56fbc36cded91d028f764f846c4f05359255cee2
Reviewed-on: https://chromium-review.googlesource.com/182452
Reviewed-by: Heng-ruey Hsu <henryhsu@chromium.org>
Commit-Queue: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>

Change-Id: I327ffd366994acac31a5672febb9ce18918c686c
Reviewed-on: https://chromium-review.googlesource.com/183307
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Gediminas Ramanauskas <gedis@chromium.org>
Commit-Queue: Gediminas Ramanauskas <gedis@chromium.org>