Add option to specify the subnet to use for host based testing
In certain network configuration for host based testing, the host has
more than one Ethernet interface. Since the host has no information on
which Ethernet interface the DUT is on, it has to scan all interfaces,
and this causes unwanted traffic on other interfaces.

To solve this, let's add an option to specify allowed subnets in a
board-specific file. If the file is absent, default to 192.168.0.*.

BUG=None
TEST=Instrument the discoverer to show which subnet it's scanning.
Change the content of /usr/local/factory/board/host_based_subnets, and
check the behavior changes accordingly.

Change-Id: If62f125bea31f69bc66adbce8ec72f30162cf827
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/218092
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>