Include test path in the post_shutdown shared datum.
In case where a shutdown test was executed and is skipped upon the
next boot, the test will never get a chance to run throuh the
post-shutdown check in the shutdown test. This means that the
post_shutdown shared datum will not be cleared and it will cause
problem to the next shutdown test.

The CLs fixes this problem by including test path in the
post_shutdown shared datum so every shutdown test will only look
for the post_shutdown tag of its own.

BUG=None
TEST=Manually on DUT.

Change-Id: Id607d6ff921e5696cd94545044cc64e229b7559b
Reviewed-on: https://chromium-review.googlesource.com/218423
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>