CHERRY-PICK: Preserve the test UI when presenter reconnects
When the connection between the presenter and the device drops, just
hide the UI instead of destroying it. Upon reconnection, compares the
UUID which is randomly generated for each UI session, and reuses the UI
from the last connection if it's the same session. This way, we can sort
of preserve the UI of the active test. However, if some events are lost
when the connection is broken, the UI might be in an inconsistent state.
Still, this change should help us in most cases.

BUG=chrome-os-partner:32670
TEST=Run a test with UI. Unplug the Ethernet cable and see the UI goes
away. Re-plug the cable and see the UI of the active test.

Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/222942
Reviewed-by: Hung-Te Lin <hungte@chromium.org>

(cherry picked from commit ac358da97ec4ea060f6b88a5de0aff7433ddbc22)

Change-Id: I07492c99f2834c66e41cd5f9647a13a102a7a6c5
Reviewed-on: https://chromium-review.googlesource.com/224560
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>