Set X env variables in goofy_control.
We used to set these X env variables in goofy processes but they are
removed recently. Set these X env variables in goofy_control so we do
not need to set them everywhere in the code.

We will need to remove dependency on X eventually due to freon. This
is just a temporary measure.

BUG=None
TEST=Manually on my DUT.

Change-Id: Ifb64086edd8d3cff86def923dfd986004b2d9cf8
Reviewed-on: https://chromium-review.googlesource.com/224603
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>