CHERRY-PICK: [Goofy] Handles NotImpelmentedError exception.
We plan to run Goofy on a test host. However, test host is not a ChromeOS
device, we want to run ChromiumOS on it with base board class, which raises
NotImplementedError for all methods by default.

Use self._can_charge to prevent logging error if the board is not chargable.

BUG=chrome-os-partner:30020
TEST=Manual test on a X86-generic test host.

Reviewed-on: https://chromium-review.googlesource.com/225047
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Dean Liao <deanliao@chromium.org>
Commit-Queue: Dean Liao <deanliao@chromium.org>
Tested-by: Dean Liao <deanliao@chromium.org>
Change-Id: Ibd9a64aadabc10b385ab5563c442a3e22e3b6194
Reviewed-on: https://chromium-review.googlesource.com/236501