CHERRY-PICK: Respond to presenter/controller pings in a separate process
The presenter and the controller ping each other periodically to
determine if the link is still good. The pings are currently processed
in the same threads the link managers are in. On slow devices, if Goofy
is busy, the pings get processed slowlier and sometimes it's too slow
that the presenter thinks the controller is down.

Avoid this by starting a standalone "ping server", which is only
responsible for responding to pings.

BUG=chrome-os-partner:33226
TEST=Run controller on Beaglebone. Doesn't see broken connection as
often.

Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/226640

(cherry picked from commit dfd893e3fbd36b1059f0dc5433ad7474f1f4a00a)

Change-Id: Iea7c26ac66237890b3cf0eae4b11fc9035dbf62d
Reviewed-on: https://chromium-review.googlesource.com/237673
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>