CHERRY-PICK: Make GoofyPresenter retry when the test UI fails to load
When the controller is just starting, there is a chance that the link
manager is up, but the web server serving the test UI isn't. If we
happen to make connection with the controller in this case, the test UI
will fail to load. When this happens, instead of doing nothing, the UI
presenter app should report the failure to GoofyPresenter, and the
GoofyPresenter can then order a retry.

BUG=chrome-os-partner:33226
TEST=unittest
TEST=Instrument the presenter to log retries. Repeatedly restart the
controller and see the presenter retry sometimes. When it retries, the
UI eventually loads.

Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/226922
Reviewed-by: Hung-Te Lin <hungte@chromium.org>

(cherry picked from commit 1f299ea9d4a44af496c03b38c64e976eae09a2cf)

Change-Id: I39839c640945d3d2f20a859344c5547068bb6ad2
Reviewed-on: https://chromium-review.googlesource.com/237674
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>