CHERRY-PICK: Registration code: raise error if user code and group code are the same
Currently we check VPD user/group bind values by their format, but not
checking their actual data. It was found that some devices were assigned
with same user/group values and that should be apparently wrong value,
according to their definition. To catch that failure, we want to add a simple
test to assert when user/group codes are the same.

This is only required for legacy format as we will check the type (user
or group) for new-style codes.

BUG=chrome-os-partner:33473
TEST=Manual
TEST=unit test

Original-Change-Id: I882b574766efa1e2b6a97394542b571195107e1b
Reviewed-on: https://chromium-review.googlesource.com/227015
Commit-Queue: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>

Change-Id: I431615d8742f3a379b7574a68de1c2d918bba2ca
Reviewed-on: https://chromium-review.googlesource.com/227292
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Commit-Queue: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>