Registration code: raise error if user code and group code are the same
Currently we check VPD user/group bind values by their format, but not
checking their actual data. It was found that some devices were assigned
with same user/group values and that should be apparently wrong value,
according to their definition. To catch that failure, we want to add a
simple test to assert when user/group codes are the same.

This is only required for legacy format as we will check the type (user
or group) for new-style codes.

BUG=chrome-os-partner:33473
TEST=Manual
TEST=unit test

Change-Id: Id1595d617e362b905d3c23e19df8e224a555a90b
Reviewed-on: https://chromium-review.googlesource.com/227293
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Commit-Queue: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>