Add sets_enabled argument to touchscreen_wrap test
touchscreen_wrap.py attempts to enable and potentially disable
the touchscreen device. It does this by shelling out to a script
/opt/google/xinput.sh. This script is not present on all devices,
and if it is missing, the function produces an error leading to a test
failure. Ultimately we will want to implement this functionality,
but until then, add an argument that prevents enabling or disabling
touchscreen devices.

BUG=none
TEST=manual

Change-Id: Ic0a5799c2832a57dab76ed70240d3e9bf779ad42
Reviewed-on: https://chromium-review.googlesource.com/231689
Commit-Queue: Peter Ammon <pca@chromium.org>
Tested-by: Peter Ammon <pca@chromium.org>
Reviewed-by: Vic Yang <victoryang@chromium.org>