Add 'no_host' argument for headless device testing
For headless devices, we are likely not able to set up a lot of test
hosts for run-in tests. Therefore, we need a way to run certain tests
without the presenter. For example, we may want to start the reboot
test and then let the device run the test by itself without the
presenter.

With this change, a single test or a test group may be marked as
'no_host=True', and this instructs goofy to skip waiting for the UI.

BUG=None
TEST=Run reboot test in chroot without a presenter.
TEST=Make sure goofy waits for the UI when it encounters a non-no_host
test after a no_host test.

Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/237690
Reviewed-by: Hung-Te Lin <hungte@chromium.org>

(cherry picked from commit a3cecf8e630d712af9322470223396bc59d6c4ae)

Change-Id: I3d91409296db36bd71935d2dda7ea6bbcbcad5b4
Reviewed-on: https://chromium-review.googlesource.com/240616
Reviewed-by: Gediminas Ramanauskas <gedis@chromium.org>
Commit-Queue: Gediminas Ramanauskas <gedis@chromium.org>
Tested-by: Gediminas Ramanauskas <gedis@chromium.org>