Run DHCP server in LinkManager
Instead of doing IP scanning with Discoverer, let's switch to runing
DHCP server in LinkManager.

More specifically, the presenter runs DHCP servers on all its unmanaged
Ethernet interfaces. When the DUT is connected to one of these
interfaces, the DUT sends DHCP request to the presenter. The presenter
issues an IP address and at the same time it knows to try connecting to
this DUT.

For this to work, on the presenter, there must be no existing DHCP
server running on the interfaces that we expect the DUT.

Also, to make sure the presenter forwards traffic from the DUT to the
shopfloor server, it's recommended that the presenter only has an
interface to the shopfloor server and one or more interfaces to the DUT,
and no other interfaces. If there are multiple managed interfaces, the
presenter randomly choose one for NAT.

BUG=chrome-os-partner:35391
TEST=Run presenter on Link with three interfaces:
       - A managed interface to my workstation.
       - Two unmanaged interfaces.
     Run controller on Spring and performs the following tests:
      1. Connects Spring to either unmanaged interface, and check the UI
         is presented. Also check Spring to workstation traffic is good.
      2. Switch to another unmanaged interface and repeat the check.
      3. Reboot Spring and the UI comes up again.
      4. Reboot Link and the UI comes up again.

Change-Id: Icdca3e3627113ebb4034eefccf011dc544c58dcd
Previous-Reviewed-on: https://chromium-review.googlesource.com/252490
(cherry picked from commit a471f70120261eda5a7883ce969506c4622adfc0)
Reviewed-on: https://chromium-review.googlesource.com/261133
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>