CHERRY-PICK: add tablet_mode test to verify lid switch status
In devices convertible that can be flipped to tablet
mode, the lid switch may be improperly shielded and
erroneously triggered, thus causing a system suspend
when converting to tablet mode.

This test asks the operator to flip the device to tablet
mode, and verifies that the lid switch has not been triggered.
Includes animated gif to explain device flipping, since it
is a little tricky to explain, especially in Chinese.

Also make some minor refactoring changes in related files.

BUG=chrome-os-partner:35226
TEST=Manually on DUT

Original-Change-Id: I8f3967ff7ce3cd81c012f7c3bde56de2f2399533

Change-Id: Ie4b06a583bf52677a896915d74f126acd7b606e4
Reviewed-on: https://chromium-review.googlesource.com/259485
Commit-Queue: Joel Kitching <kitching@google.com>
Tested-by: Joel Kitching <kitching@google.com>
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>