touchscreen_wrap: Check moving direction
On conducting the God's touch test in OQC, circles are supposed to show up
exactly under the touching finger. If this is not the case, the touchscreen
is considered bad. It is desirable to catch the mis-location problem too in
this test. Such a bad panel may be caught in this test when a finger moves
in some direction, its reported coordinates jump in the other directions
when the finger moves near around the problematic spot of the touchscreen.

BUG=chrome-os-partner:39377
TEST=Manually tested on a failed DUT.

Change-Id: I3673d7da7aa8a5ee9285d3397007579689a0e4e8
Reviewed-on: https://chromium-review.googlesource.com/268982
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>