overlord: factoryinstall: add timeout of DUT boot up.
From powering up the DUT to entering the kernel is almost 30 seconds,
the responding time is too long for the factory. The operator should
wait 30 secs to see the status become yellow, then do the next thing.
We use a workaround way to improve the issue.

- The operator power up the DUT
- The operator click the status screen at the overlord, and the
  status will become yellow
- If DUT doesn't successfully boot up and connect to overlord in
  40 seconds, then the status become red.

The operator can be notified the fail by the timeout, instead of waiting
by himself.

Also some bugs fixed.
- Moved the global variable window.locked into the App class.
- Do not use the status of websocket to determine the connection state,
  only use logcat event.

TEST=manually
BUG=chrome-os-partner:39839

Change-Id: I4757b46d84ab4c930c787e3976b6b687402fe517
Reviewed-on: https://chromium-review.googlesource.com/269988
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>