CHERRY-PICK: factory_bug: fix error on device having no EC
Currently we check the existence of 'ectool' to determine if a device
has EC or not. However, this is inaccurate as 'ectool' will be included
in a test image regardless of hardware status.

Use the return code of `ectool version` to determine whether to include
EC-related logs or not.

BUG=none
TEST=Manual on a EC-less device

Change-Id: Ic75a76c0416228e40040a63a0165110047bdf540
Original-Change-Id: I8de0b5633900116e84065796419985f53dc2f533
Reviewed-on: https://chromium-review.googlesource.com/273983
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Kaiyen Chang <kaiyen.chang@intel.com>
Commit-Queue: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/290220
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>