CHERRY-PICK: Overlord: factoryinstall: Store the status in cookies.
The order of machine ID is stored into cookies when the panel is locked.
Therefore we don't need to setup the order again if the operator closes
the overlord front-end accidently.
The timeout of bootup is also stored into cookies, and increase the default
value to 60 seconds.

BUG=none
TEST=Manually test factory install app

Change-Id: I3ceb1a2acb3dc43deb041f864cf689e73cefb900
Original-Change-Id: Ia74ba2f3618126ac05ed35e96fefd6c22794bd93
Reviewed-on: https://chromium-review.googlesource.com/289090
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294005
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>