CHERRY-PICK: Pass the test instance to OnTestFailure function.
For the partner who needs the test information when a test failed, we pass the
test instance to the OnTestFailure function.
Also changed the default action of OnSummaryBad not to call OnTestFailure to
prevent the function is called twice.

BUG=chrome-os-partner:43356
TEST=Manually test at device which is called OnTestFailure.

Original-Change-Id: Idc2b58a14adacff7e45af8288dd24db36dee2e13
Reviewed-on: https://chromium-review.googlesource.com/289041
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>

Conflicts:
	py/goofy/goofy.py

Change-Id: Ia97af325b344e5b993cf290a0f8ed6b73635dcbc
Reviewed-on: https://chromium-review.googlesource.com/289753
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>