wireless_radiotap: specify the D-Bus type for each element
We have switched to new D-Bus libraries and org.freedesktop.dbus.introspectable
interface is not implemented, results in this pytest not able to
correctly convert variant dictionary parameters. So the dictionary
here would be treated as an array of strings (since the first element
is a string) when passing to over to D-Bus.

Fix this by specifying the D-Bus type for each element

BUG=chrome-os-partner:45841
TEST=pass the wireless_radiotap item without D-BUS errors.

Change-Id: I351d8b8fb04f5aba8d1db0f66a9aeb36b969fdd3
Reviewed-on: https://chromium-review.googlesource.com/303233
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Commit-Queue: Kevin Cheng <kevin.cheng@intel.com>
Tested-by: Kevin Cheng <kevin.cheng@intel.com>