goofy: Fix touch events in UI (for touchscreen_wrap).
The last missing piece of init-based Chrome startup migration.

Touchscreen_wrap has been broken in monolithic mode since the init-based
factory startup flow has been applied. Seems like for the first startup
window, if not empty, Chrome needs some explicit instruction to know if
touch events should be loaded or not.

This was not a problem in presenter mode because the presenter did an
explicit reload to get right permissions (which also enables the touch
events for most devices). This was also not a problem for OOBE or normal
boot path because they usually have their own permission and policy, and
the "first" windows is not what you saw on screen.

To allow testing touch screen without hacking opened windows, we want to
add the --touch-events to chrome_dev.conf so every started factory
instance would have right touch events in the first window.

BUG=chrome-os-partner:30239;chrome-os-partner:50847
TEST=Manually executed touchscreen_wrap for both monolithic and split mode.

Change-Id: I9a49a243623b8e7c9cdd23cf994f20828b0afaff
Reviewed-on: https://chromium-review.googlesource.com/337440
Tested-by: Michael H Yang <michael.h.yang@intel.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Michael H Yang <michael.h.yang@intel.com>