goofy: invocation: fix failure when resuming tests.
When Goofy is trying to resume tests (e.g., reboot test), it will fail
after the device is rebooted, because the dargs is not properly passed
to "invoke_pytest".

In CL:362740, we move the timing to resolve dargs before calling
invoke_pytest, and the resolved dargs should be passed to invoke_pytest.
However, when we are resuming the test, the variable that should be
passed to invoke_pytest is not initialized.

In this CL, read the resolved dargs from metadata and pass it to
invoke_pytest.

BUG=None
TEST=manual

Original-Change-Id: I014dc85a61ae32d4a15599d10401cd9c81ff55d4
Original-Reviewed-on: https://chromium-review.googlesource.com/367030
(cherry picked from commit c470d99880767d3642be64f0d1da41b673b20056)

Change-Id: I5a1055f4a4f1f76883613a698f01c084fa8d65ea
Reviewed-on: https://chromium-review.googlesource.com/381035
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>