CHERRY-PICK: finalize: start response listener right away
The response listener was started *after* gooftool starts wiping.
However, we always try to shutdown response listener in tearDown, this
would block pytest if gooftool failed to start wiping because the
response listener is not running and nothing can be shutdown.

In this CL, we start the response listener in a daemon thread right
after it is created, so the teardown function would always work
properly.

BUG=None
TEST=manual

Change-Id: If47e1d058d017812b946b1ec6efc3e03cf2356d4
Original-Change-Id: I023ee5728f300a63f598f06e1a0b853ca5e9304c
Reviewed-on: https://chromium-review.googlesource.com/390617
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Commit-Queue: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>