gale: ignore error when enabling EC write protect
When enabling EC write protection, flashrom(8) might fail because
setting RO_NOW is not supported.
In this CL, if flashrom(8) return none zero value, we will parse
standard error, if it's because RO_NOW is not set but RO_AT_BOOT is set
successfully, we will ignore the error.

BUG=None
TEST=None

Change-Id: Icfd3d680dafd825607060e657826d9d909eda2ea
Reviewed-on: https://chromium-review.googlesource.com/398905
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>