power: Fix GetBatteryCurrent on sysfs to return in correct sign
In the past, the value of current_now file in sysfs is always
non-negative. But on some new devices, such as Gru and Kevin, it will be
negative when discharging, and make GetBatteryCurrent return positive
values most of the time. Should ignore the sign of current_now.

BUG=none
TEST=manually

Change-Id: Ibd6fd7c9a07b0b694207ce8428bbfee0cbf8be7c
Reviewed-on: https://chromium-review.googlesource.com/415748
Reviewed-by: Youcheng Syu <youcheng@google.com>
Commit-Queue: Youcheng Syu <youcheng@chromium.org>
Tested-by: Youcheng Syu <youcheng@chromium.org>