test: rework pytest display_point challenge response
The old way to respond to challenge in display_point pytest is using a
text input, which would need a virtual keyboard to be popped out if we
want to use touchscreen only. But the virtual keyboard would interfere
with the following fullscreen test.

Since we probably never want to have challenge of more than 10 points
(default is 3), a single number keypress is enough for challenge
response. This makes using touchscreen only easier.

This also simplify the process of answering. Assuming that the answers
are 2 and 3.
Keystrokes to pass old test:
  [Space] [Space] 2 [Enter] [Space] 3 [Enter]
Keystrokes to pass new test:
  [Space] 2 3

BUG=chromium:670277
TEST=make test
TEST=manually tested, put the test on DUT, and check that the test can
     be easily completed with touchscreen only.

Change-Id: Ia3bfbfa2812572a6c638940af0ba747e14307788
Reviewed-on: https://chromium-review.googlesource.com/417683
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/441689
Reviewed-by: Youcheng Syu <youcheng@google.com>
Commit-Queue: Youcheng Syu <youcheng@google.com>
Tested-by: Youcheng Syu <youcheng@google.com>