CHERRY-PICK: hwid: Change 'mandatory rules' to 'default rules'.
The 'mandatory rules' were created by merging what most Chromebook
projects would do into HWID standard flow. However, we do see exceptions
in future for example ECHO codes may be not mandatory on ChromeOS Code
devices, or Zerg devices that need to have ECHO code device different
from board names.

As a result, we believe these 'mandatory' rules should be something that
can be disabled or overridden. An alternative is to do that in
config_utils, but the concern is these are usually what Google wants to
control so it is better to be in HWID database file, which should be
reviewed and released by Google.

As a result and compared with all existing HWID files, it seems
reasonable to change the 'mandatory' to 'default', and only execute them
if the HWID database file does not specify these rules.

To override, define your own rule:

 - name: verify.vpd.ro
 - evaluate: ...

 - name: verify.vpd.rw
 - evaluate: ...

BUG=chrome-os-partner:61690
TEST=make test

Change-Id: Icdfe4b484b412eb0e731205d641eb87eeaa7d253
Reviewed-on: https://chromium-review.googlesource.com/446796
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>