CHERRY-PICK: umpire: Change dynamic shopfloor handler to static Shopfloor Service API.
Umpire designed supporting shopfloor handlers in a very dynamic way, to
help different clients being able to connect to the associated handlers.

However, we find that introducing more troubles because:
 - It becomes impossible to "lock" shopfloor code when we need to deal
   with lots of local changes in toolkit across many developers.
 - Umpire now runs inside Docker and the dependency of board handlers
   may break easily, from Python package to port & directory binding.
 - The shopfloor handlers don't work pretty well when we update package
   in place.
 - Running multiple handlers under FastCGI also made syncing more
   difficult.

Comparing with what's needed in practice, it is believed that calls to
shopfloor backends should be isolated, defined with clear interface,
and no difference for devices in same build. If there will be any
special rule, it should be implemented directly on server side, not the
handler proxy.

BUG=chromium:640162
TEST=make test; ./setup/cros_docker.sh build; cros_docker.sh umpire test
     installed cros_docker, start Dome, created an Umpire session.

Change-Id: Ib2f8e4876bbd7375a1f19db91f1006703050fa4b
Original-Change-Id: If485a19e46bae3dfd24cb57eb87302bded4b2fa4
Reviewed-on: https://chromium-review.googlesource.com/537273
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/542158
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Commit-Queue: Wei-Han Chen <stimim@chromium.org>