CHERRY-PICK: overlord: Fix a small race condition.
The socket IO server in Overlord is used in both the http server and
socket server. Separate the initialize step with the run-server step, so
we can guarantee servers are properly initialized before it is used.

This is the correct fix for CL:471146.

BUG=None
TEST=make test
TEST=run overlordd and open the web interface many times, check that
     there's no runtime error for overlordd.

Change-Id: Ie54041d42354f7bdb673fbd7620d4446baedaf61
Original-Change-Id: I504555ac5f0b66a21a75c8397a913784e566d2d9
Reviewed-on: https://chromium-review.googlesource.com/540979
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/542164
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Commit-Queue: Wei-Han Chen <stimim@chromium.org>