hwid: Refactor in py/hwid/v3/ to prepare for extending rule.Value
A value in hwid database could only be either raw string or regular
expression.  But during the factory build, we saw that the match rule of
some hardware attribute such as battery capacity need to be "larger than or
equal to a specific number".  So we need to change `rule.Value` class to be
extendable for future requirement.

BUG=None
TEST=run all unittest under py/hwid/v3

Change-Id: I288f1eaf17732d9518adf7218636977e9d5b7910
Reviewed-on: https://chromium-review.googlesource.com/544370
Tested-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>