wifi-throughput: extend timeout and transmission time
When invoking the iperf client on DUT, the default timeout to
connect to the station is 5 seconds. In some network environment,
it might not be long enough. This commit changes it to 10 seconds.

For the transmission time, this commit increase it from 5 seconds
to 10 seconds. The test use iperf to send TCP/IP packets for a
specified duration. In some network environment, the transfer speed
is quite slow in the first 5 seconds. This might due to the slow-start
property of TCP/IP protocol.
An alternative option is to send UDP datagrams in our test. But, since
TCP/IP is still an important protocol, it might be a good idea to stick
on testing TCP/IP.

BUG=none
TEST=none

Change-Id: Iea68829d155e6111b002dc7fb79eda4e170c7829
Reviewed-on: https://chromium-review.googlesource.com/556439
Commit-Queue: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>