factory: Add value range check for sampling ALS value.
We found that the sampling ALS value from the fixture may be unstable.
This could be caused by the unstable electric environment or fixture's
unstability after turning light settling. To solve that, we should detect
the outlier value when sampling lux and raise error if the value is
outside expected range.

TEST=none
BUG=b:67075656

Change-Id: I2d121edcb0f5d481af6e539b5cd65c657e19e980
Reviewed-on: https://chromium-review.googlesource.com/691676
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Yilun Lin <yllin@chromium.org>
Tested-by: Yilun Lin <yllin@chromium.org>