umpire proxy: don't log network connection issue
Device is not always online in factory.  If a device is running pytest
like sync_shopfloor without a network connection (e.g. operators go to
lunch), it might accumulate thousands of same network error.  This would
cause a lot of trouble in logging system.  Let's stop logging this for
now as a workaround (a proper fix is introduced in ToT)

BUG=None
TEST=None

Change-Id: Iffbbbac363347cbdaa6776f63ae3ec067f547939
Reviewed-on: https://chromium-review.googlesource.com/750881
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Commit-Queue: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>