test list iterator: fix CheckSkip for parallel group
When a group is parallel, it will be considered 'IsRunnableTest'.
However, when we are checking if we need to rerun a test that was
'SKIPPED', the rule should be:

* For test groups, they need retest if their run_if are set, and
  evaluate to True.
* For **leaf tests**, they need retest if their run_if are not set, or
  evaluate to True.

Because a **runnable test** might be a test group.

BUG=chromium:796485
TEST=make test

Change-Id: I8b77ad5957374b1632882deacd47241e3e836318
Reviewed-on: https://chromium-review.googlesource.com/838580
Reviewed-by: Youcheng Syu <youcheng@chromium.org>
Commit-Queue: Youcheng Syu <youcheng@chromium.org>
Tested-by: Youcheng Syu <youcheng@chromium.org>
Trybot-Ready: Youcheng Syu <youcheng@chromium.org>