goofy: Change the init order to prevent 404 not found error.
Change the initialization order of Goofy, so that the files needed for
the initial index.html to load is registered before the Goofy server
come up. Chrome seems to automatically retry connection if it get
connection refused (That is, before Goofy come up), but would show an
error page if it get 404, so we want to prevent the latter case.

BUG=b:74049291
TEST=make test

Change-Id: I8989801d4dbfd2a9877559df143d7ddc4f847d53
Reviewed-on: https://chromium-review.googlesource.com/948124
Reviewed-by: Youcheng Syu <youcheng@chromium.org>
Commit-Queue: Youcheng Syu <youcheng@chromium.org>
Tested-by: Youcheng Syu <youcheng@chromium.org>
Trybot-Ready: Youcheng Syu <youcheng@chromium.org>