camera: Send image back through Goofy RPC instead of event.
Since the event server does broadcast on all data received, and has an
upper limit on the size of each message, it's not suitable for passing
large blob. Change how the image is sent back to pytest to using
goofy_rpc to put the content to a temporary file.

This speed up the transmitting time for each frame from 0.2 seconds to
0.1 seconds in my testing.

BUG=None
TEST=make test
TEST=manually on DUT.

Change-Id: Ia05df2e60a1cc0b794f89a81d6bbe0166a6a829e
Reviewed-on: https://chromium-review.googlesource.com/987659
Reviewed-by: Youcheng Syu <youcheng@chromium.org>
Commit-Queue: Youcheng Syu <youcheng@chromium.org>
Tested-by: Youcheng Syu <youcheng@chromium.org>
Trybot-Ready: Youcheng Syu <youcheng@chromium.org>