Makefile: Revise resource system to a more easily customized way.
The factory-board and resource system has enabled Chromium OS factory
software to be easily customized. But to add files to the factory
related packages, for example chromeos-base/factory_installer or netboot
initramfs, we have to create a dedicated ebuild file and deal with the
portage virtual package system, which is pretty difficult for partners.

As a result, we want to generalize the resource system so all factory
related customization can be done in a single factory-board package.

The new folder 'resources' and the *.rsrc files are the new way to
construct resource tarball. Each *.rsrc contains the declaration of a
resource file - either from SYSROOT or BOARD_FILES_DIR or public repo.

BUG=chromium:837561
TEST=make test; make resource

Change-Id: Iaa884381d7928f343ba8b5694bfdaac0e9c1cf8b
Reviewed-on: https://chromium-review.googlesource.com/1034512
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Trybot-Ready: Hung-Te Lin <hungte@chromium.org>