factory_bug: support saving logs when booting from USB
Eve RMA is using USB image (booting from USB disk installed with test
image + toolkit), however, factory_bug would behave differently when
we are booting from a USB disk.  It will try to mount the hard drive and
copy logs on the hard drive to the USB disk.  This is not what we want
for USB RMA process.

This CL change the logic, so if '--mount' is not specified, we will
still save logs on USB disk (to output_dir).

BUG=None
TEST=manually

Change-Id: I07dfa35b9e6319ecbccab8ccc061937626b71924
Reviewed-on: https://chromium-review.googlesource.com/1059005
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Commit-Queue: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>